refactor: 예외 패키지를 handler/code/type으로 분리
