+++ src/main/java/com/chukchuk/haksa/domain/auth/service/RefreshTokenService.java
@@ -5,9 +5,9 @@ import com.chukchuk.haksa.domain.auth.entity.RefreshToken;
 import com.chukchuk.haksa.domain.auth.repository.RefreshTokenRepository;
 import com.chukchuk.haksa.domain.user.model.User;
 import com.chukchuk.haksa.domain.user.repository.UserRepository;
-import com.chukchuk.haksa.global.exception.EntityNotFoundException;
-import com.chukchuk.haksa.global.exception.ErrorCode;
-import com.chukchuk.haksa.global.exception.TokenException;
+import com.chukchuk.haksa.global.exception.type.EntityNotFoundException;
+import com.chukchuk.haksa.global.exception.code.ErrorCode;
+import com.chukchuk.haksa.global.exception.type.TokenException;
 import com.chukchuk.haksa.global.logging.annotation.LogTime;
 import com.chukchuk.haksa.global.security.service.JwtProvider;
 import io.jsonwebtoken.Claims;
