+++ src/main/java/com/chukchuk/haksa/domain/student/model/GradeType.java
@@ -1,7 +1,7 @@
 package com.chukchuk.haksa.domain.student.model;
 
-import com.chukchuk.haksa.global.exception.CommonException;
-import com.chukchuk.haksa.global.exception.ErrorCode;
+import com.chukchuk.haksa.global.exception.type.CommonException;
+import com.chukchuk.haksa.global.exception.code.ErrorCode;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
