+++ src/main/java/com/chukchuk/haksa/global/exception/type/CommonException.java
@@ -1,5 +1,7 @@
-package com.chukchuk.haksa.global.exception;
+package com.chukchuk.haksa.global.exception.type;
 
+
+import com.chukchuk.haksa.global.exception.code.ErrorCode;
 
 public class CommonException extends BaseException {
     public CommonException(ErrorCode errorCode) {
