+++ src/main/java/com/chukchuk/haksa/global/exception/type/EntityNotFoundException.java
@@ -1,4 +1,6 @@
-package com.chukchuk.haksa.global.exception;
+package com.chukchuk.haksa.global.exception.type;
+
+import com.chukchuk.haksa.global.exception.code.ErrorCode;
 
 public class EntityNotFoundException extends BaseException {
     public EntityNotFoundException(ErrorCode errorCode) {
