+++ src/main/java/com/chukchuk/haksa/global/exception/type/TokenException.java
@@ -1,4 +1,6 @@
-package com.chukchuk.haksa.global.exception;
+package com.chukchuk.haksa.global.exception.type;
+
+import com.chukchuk.haksa.global.exception.code.ErrorCode;
 
 public class TokenException extends BaseException {
     public TokenException(ErrorCode errorCode) {
